Compressed all globals into a single closure
All globals originally at top are now in a closure called fobj. Along
with this, a lot of the functions which were outside have been put in
this single large closure.

TODO:
If I make fobj local then all functions need to take a parameter fobj.
This will interfere with syntax in HTML where functions are called.