Moving towards literate programming
Have created a vim plugin to help in literate programming. The plugin
is in my Dropbox/vimscripts and it's called
markdown_code_extractor.vim. There are still some bugs inside the
function. The most scary bug is that when there is an error in the js
file, the entire Markdown gets deleted. I had to refresh the source
from Dropbox version control.